+++ apps/nextjs-website/src/app/privacy-policy/page.tsx
@@ -2,6 +2,7 @@ import PrivacyPolicyTemplate from '@/components/templates/PrivacyPolicyTemplate/
 import { baseUrl } from '@/config';
 import { makeMetadata } from '@/helpers/metadata.helpers';
 import { Metadata } from 'next';
+import ContentWrapper from '@/components/atoms/ContentWrapper/ContentWrapper';
 
 export async function generateMetadata(): Promise<Metadata> {
   return makeMetadata({
@@ -12,7 +13,11 @@ export async function generateMetadata(): Promise<Metadata> {
 }
 
 const PrivacyPolicy = () => {
-  return <PrivacyPolicyTemplate />;
+  return (
+    <ContentWrapper>
+      <PrivacyPolicyTemplate />
+    </ContentWrapper>
+  );
 };
 
 export default PrivacyPolicy;