+++ apps/nextjs-website/src/lib/strapi/makeProps/makeGuideListPages.ts
@@ -5,10 +5,10 @@ import { makeBaseProductWithoutLogoProps } from './makeProducts';
 import { GuideCardProps } from '@/components/molecules/GuideCard/GuideCard';
 import { BaseGuide } from '@/lib/strapi/types/guide';
 import _ from 'lodash';
-import { StrapiGuideListPaginated } from '@/lib/strapi/types/guideList';
+import { StrapiGuideListPages } from '@/lib/strapi/types/guideListPage';
 
 export function makeGuideListPagesProps(
-  strapiGuideListPages: StrapiGuideListPaginated
+  strapiGuideListPages: StrapiGuideListPages
 ): readonly GuideListPageProps[] {
   return strapiGuideListPages.data.map(({ attributes }) => {
     const product = makeBaseProductWithoutLogoProps(attributes.product.data);
@@ -34,7 +34,7 @@ export function makeGuideListPagesProps(
           : attributes.product.data.attributes.bannerLinks?.map(
               makeBannerLinkProps
             ),
-      seo: attributes.seo,
+      seo: attributes.seo || undefined,
       updatedAt: attributes.updatedAt,
     } satisfies GuideListPageProps;
   });